api/requirements.txt
fastapi
mangum
python-dotenv
requests